+++ CesiumGeometry/test/TestTransforms.cpp
@@ -1,9 +1,12 @@
+#include "CesiumGeometry/Axis.h"
 #include "CesiumGeometry/Transforms.h"
 
 #include <CesiumUtility/Math.h>
 
 #include <catch2/catch.hpp>
 #include <glm/mat4x4.hpp>
+
+using namespace CesiumGeometry;
 
 TEST_CASE("Transforms convert the axes correctly") {
 
@@ -12,35 +15,102 @@ TEST_CASE("Transforms convert the axes correctly") {
   glm::dvec4 Z_AXIS{0.0, 0.0, 1.0, 0.0};
 
   SECTION("Y_UP_TO_Z_UP transforms  X to  X,  Y to -Z, and  Z to Y") {
-    REQUIRE(X_AXIS * CesiumGeometry::Transforms::Y_UP_TO_Z_UP == X_AXIS);
-    REQUIRE(Y_AXIS * CesiumGeometry::Transforms::Y_UP_TO_Z_UP == -Z_AXIS);
-    REQUIRE(Z_AXIS * CesiumGeometry::Transforms::Y_UP_TO_Z_UP == Y_AXIS);
+    REQUIRE(X_AXIS * Transforms::Y_UP_TO_Z_UP == X_AXIS);
+    REQUIRE(Y_AXIS * Transforms::Y_UP_TO_Z_UP == -Z_AXIS);
+    REQUIRE(Z_AXIS * Transforms::Y_UP_TO_Z_UP == Y_AXIS);
   }
   SECTION("Z_UP_TO_Y_UP transforms  X to  X,  Y to  Z, and  Z to -Y") {
-    REQUIRE(X_AXIS * CesiumGeometry::Transforms::Z_UP_TO_Y_UP == X_AXIS);
-    REQUIRE(Y_AXIS * CesiumGeometry::Transforms::Z_UP_TO_Y_UP == Z_AXIS);
-    REQUIRE(Z_AXIS * CesiumGeometry::Transforms::Z_UP_TO_Y_UP == -Y_AXIS);
+    REQUIRE(X_AXIS * Transforms::Z_UP_TO_Y_UP == X_AXIS);
+    REQUIRE(Y_AXIS * Transforms::Z_UP_TO_Y_UP == Z_AXIS);
+    REQUIRE(Z_AXIS * Transforms::Z_UP_TO_Y_UP == -Y_AXIS);
   }
 
   SECTION("X_UP_TO_Z_UP transforms  X to -Z,  Y to  Y, and  Z to  X") {
-    REQUIRE(X_AXIS * CesiumGeometry::Transforms::X_UP_TO_Z_UP == -Z_AXIS);
-    REQUIRE(Y_AXIS * CesiumGeometry::Transforms::X_UP_TO_Z_UP == Y_AXIS);
-    REQUIRE(Z_AXIS * CesiumGeometry::Transforms::X_UP_TO_Z_UP == X_AXIS);
+    REQUIRE(X_AXIS * Transforms::X_UP_TO_Z_UP == -Z_AXIS);
+    REQUIRE(Y_AXIS * Transforms::X_UP_TO_Z_UP == Y_AXIS);
+    REQUIRE(Z_AXIS * Transforms::X_UP_TO_Z_UP == X_AXIS);
   }
   SECTION("Z_UP_TO_X_UP transforms  X to  Z,  Y to  Y, and  Z to -X") {
-    REQUIRE(X_AXIS * CesiumGeometry::Transforms::Z_UP_TO_X_UP == Z_AXIS);
-    REQUIRE(Y_AXIS * CesiumGeometry::Transforms::Z_UP_TO_X_UP == Y_AXIS);
-    REQUIRE(Z_AXIS * CesiumGeometry::Transforms::Z_UP_TO_X_UP == -X_AXIS);
+    REQUIRE(X_AXIS * Transforms::Z_UP_TO_X_UP == Z_AXIS);
+    REQUIRE(Y_AXIS * Transforms::Z_UP_TO_X_UP == Y_AXIS);
+    REQUIRE(Z_AXIS * Transforms::Z_UP_TO_X_UP == -X_AXIS);
   }
 
   SECTION("X_UP_TO_Y_UP transforms  X to -Y,  Y to  X, and  Z to  Z") {
-    REQUIRE(X_AXIS * CesiumGeometry::Transforms::X_UP_TO_Y_UP == -Y_AXIS);
-    REQUIRE(Y_AXIS * CesiumGeometry::Transforms::X_UP_TO_Y_UP == X_AXIS);
-    REQUIRE(Z_AXIS * CesiumGeometry::Transforms::X_UP_TO_Y_UP == Z_AXIS);
+    REQUIRE(X_AXIS * Transforms::X_UP_TO_Y_UP == -Y_AXIS);
+    REQUIRE(Y_AXIS * Transforms::X_UP_TO_Y_UP == X_AXIS);
+    REQUIRE(Z_AXIS * Transforms::X_UP_TO_Y_UP == Z_AXIS);
   }
   SECTION("Y_UP_TO_X_UP transforms  X to  Y,  Y to -X, and  Z to  Z") {
-    REQUIRE(X_AXIS * CesiumGeometry::Transforms::Y_UP_TO_X_UP == Y_AXIS);
-    REQUIRE(Y_AXIS * CesiumGeometry::Transforms::Y_UP_TO_X_UP == -X_AXIS);
-    REQUIRE(Z_AXIS * CesiumGeometry::Transforms::Y_UP_TO_X_UP == Z_AXIS);
+    REQUIRE(X_AXIS * Transforms::Y_UP_TO_X_UP == Y_AXIS);
+    REQUIRE(Y_AXIS * Transforms::Y_UP_TO_X_UP == -X_AXIS);
+    REQUIRE(Z_AXIS * Transforms::Y_UP_TO_X_UP == Z_AXIS);
+  }
+}
+
+TEST_CASE("Gets up axis transform") {
+  const glm::dmat4 Identity(1.0);
+
+  SECTION("Gets X-up to X-up transform") {
+    CHECK(
+        Transforms::getUpAxisTransform(
+            CesiumGeometry::Axis::X,
+            CesiumGeometry::Axis::X) == Identity);
+  }
+
+  SECTION("Gets X-up to Y-up transform") {
+    CHECK(
+        Transforms::getUpAxisTransform(
+            CesiumGeometry::Axis::X,
+            CesiumGeometry::Axis::Y) == Transforms::X_UP_TO_Y_UP);
+  }
+
+  SECTION("Gets X-up to Z-up transform") {
+    CHECK(
+        Transforms::getUpAxisTransform(
+            CesiumGeometry::Axis::X,
+            CesiumGeometry::Axis::Z) == Transforms::X_UP_TO_Z_UP);
+  }
+
+  SECTION("Gets Y-up to X-up transform") {
+    CHECK(
+        Transforms::getUpAxisTransform(
+            CesiumGeometry::Axis::Y,
+            CesiumGeometry::Axis::X) == Transforms::Y_UP_TO_X_UP);
+  }
+
+  SECTION("Gets Y-up to Y-up transform") {
+    CHECK(
+        Transforms::getUpAxisTransform(
+            CesiumGeometry::Axis::Y,
+            CesiumGeometry::Axis::Y) == Identity);
+  }
+
+  SECTION("Gets Y-up to Z-up transform") {
+    CHECK(
+        Transforms::getUpAxisTransform(
+            CesiumGeometry::Axis::Y,
+            CesiumGeometry::Axis::Z) == Transforms::Y_UP_TO_Z_UP);
+  }
+
+  SECTION("Gets Z-up to X-up transform") {
+    CHECK(
+        Transforms::getUpAxisTransform(
+            CesiumGeometry::Axis::Z,
+            CesiumGeometry::Axis::X) == Transforms::Z_UP_TO_X_UP);
+  }
+
+  SECTION("Gets Z-up to Y-up transform") {
+    CHECK(
+        Transforms::getUpAxisTransform(
+            CesiumGeometry::Axis::Z,
+            CesiumGeometry::Axis::Y) == Transforms::Z_UP_TO_Y_UP);
+  }
+
+  SECTION("Gets Z-up to Z-up transform") {
+    CHECK(
+        Transforms::getUpAxisTransform(
+            CesiumGeometry::Axis::Z,
+            CesiumGeometry::Axis::Z) == Identity);
   }
 }