+++ CHANGES.md
@@ -11,6 +11,10 @@
 
 - Added `CesiumUtility::gzip`.
 
+##### Additions :tada:
+
+- Added `CesiumGeometry::Transforms::getUpAxisTransform` to get the transform that converts from one up axis to another.
+
 ##### Fixes :wrench:
 
 - Errors while loading raster overlays are now logged. Previously, they were silently ignored in many cases.